feat: add Game and User Info
This adds GameInfo and UserInfo jsons to prepare for the recording board,
which use some global state to determine the scores should be recorded.

GameInfo contains game titles and the current phase and stage of the game.
UserInfo contains basic user information.
The user's group would be considered by recording board as well as the current phase and stage.